Add basic auth filter stub
This patch adds basic auth filter stub. Yes, the first filter that we
want to build is a basic auth filter.

From the spec[1], we know that basic auth ultimately deals with the request
from the client (downstream). If that particular request has the
necessary authorization header, the proxy (in our case is envoy) would
'forward' that request to the upstream (a server sits behind envoy).

  client -> [request] -> (decoder) ->
                                     \
                                      upstream
                                     /
 client <- [response] <- (encoder) <-

The decoder and encoder are inside envoy.

As we can get from [2], one way to extend envoy is via filters. In this
case we are interested in HTTP filters as specified in [3].

For the basic auth specific case, we need to build a decoder filter. The
building blocks to build a decoder filter is to extend the
Http::StreamDecoderFilter. This commit exhibits a simple class, i.e.
BasicAuthFilter that implements Http::StreamDecoderFilter that
forwards any requests. At this point, there is also a unit test, in
which we interested in making sure the filter_->decoderHeaders(_, _)
returns Http::FilterHeadersStatus::Continue.

Our logic definitely will be in the filter_->decodeHeaders(_, _) only,
since basic auth as specified in [1] worries nothing but request headers.

Refs:
- [1] https://tools.ietf.org/html/rfc7617
- [2] https://www.envoyproxy.io/docs/envoy/latest/extending/extending
- [3] https://www.envoyproxy.io/docs/envoy/latest/intro/arch_overview/http_filters#arch-overview-http-filters

Signed-off-by: Dhi Aurrahman <dio@rockybars.com>