Add basic auth implementation
This patch contains several things, and this is the main thing of the basic
auth implementation.

Firstly, we introduce a config struct. This config struct holds:

- username
- password
- realm
- message

username and password are joined as a string separated by ':', i.e.
username:password then it is encoded as a base64 string. The encoded
value will be compared to the value of the request authorization header,
i.e. Basic <base64>. If it is matched, it allows the request to be conveyed to
the upstream. If it is not matched, the proxy sends back 401 response
with a header entry WWW-Authenticate: Basic <realm>, plus response body
with <message> as it's value. The content-type of it is text/plain,
surely it can be improved to override the content-type.

Secondly, we have a private function authenticated(_) that does checking
of the request headers against the configuration.

Thirdly, we prepare test helpers for unauthorized requests and
the authorized one.

Signed-off-by: Dhi Aurrahman <dio@rockybars.com>