Add configuration factory and e2e test
To integrate our basic auth filter into the envoy binary, we need to have a
filter config factory. This is responsible for creating the filter
instance. The factory, normally is fed by a specified config in proto
format. However, at this point, we hard-coded it. We will specify our filter
config in proto in the next commit.

We also need to register the factory to "envoy registry", letting envoy
to know how to create an instance of the filter. The registartion is
done by name.

This patch also provides an end-to-end test for a HTTP decoder filter.
It does basically the same as filter unit testing but via envoy main
process.

Signed-off-by: Dhi Aurrahman <dio@rockybars.com>